requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
drf-yasg>=1.21.0
psycopg2-binary>=2.9.6
django-cors-headers>=4.0.0
